chore: preact-signals api
Reimplement the API of Preact Signals using current understanding of
signals.

Signal
Implements the signal flow, once initialized `const sig = Signal(0)`
setting the value is `sig.value = 1` and retrieving the value through
`sig.value`

If a signals value is retrieved in callback sent to effect, or computed,
the callback will be called every time the signal value is updated.

Batch was tricky to wrap my head around, I kept trying to put it with
the signal getter method, but batch is probably setting, so it should
be with the setters 🤦. Uses simple depth tracking to allow for nested
batches to be triggered. Then runs all changes when reaching the end of
the root batch callback.

`untracked` seems uninteresting, but it will prevent an infinite loop
when changing a signal from within itself.